Parse xml 2
Your task here is to extract data from xml on authors of an article
and add it to a list, one item for an author.
See the provided data structure for the expected format.
The tags for first name, surname and email should map directly
to the dictionary keys, but you have to extract the attributes from the "insr" tag
and add them to the list for the dictionary key "insr"